Added: memo, double click to conversations, show new conversation, component Meesage.js
